moved test files around so it's a bit cleaner
