+++ test/scripts/initDb.ts
@@ -2,7 +2,7 @@ import * as fs from 'fs'
 import { promisify } from 'util'
 
 import * as mysql from 'mysql2/promise'
-import testConfig from './testConfig'
+import testConfig from '../testConfig'
 
 const readFile = promisify(fs.readFile)
 
@@ -14,8 +14,8 @@ beforeAll(async () => {
         multipleStatements: true,
     })
 
-    const schema = await readFile(`${__dirname}/schema.sql`, 'utf8')
-    const data = await readFile(`${__dirname}/data.sql`, 'utf8')
+    const schema = await readFile(`${__dirname}/../fixtures/schema.sql`, 'utf8')
+    const data = await readFile(`${__dirname}/../fixtures/data.sql`, 'utf8')
 
     await conn.query(schema)
     await conn.query(data)
